pause publishing while reconnecting
